+++ src/main/java/com/barlinc/unusual_prehistory/blocks/plant/TallAmbientPlantBlock.java
@@ -1,135 +1,64 @@
 package com.barlinc.unusual_prehistory.blocks.plant;
 
+import com.barlinc.unusual_prehistory.entity.mob.base.AmbientMob;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.util.Mth;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.context.BlockPlaceContext;
-import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.entity.MobSpawnType;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.LevelAccessor;
-import net.minecraft.world.level.LevelReader;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.StateDefinition;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.block.state.properties.DoubleBlockHalf;
-import net.minecraft.world.level.block.state.properties.EnumProperty;
-import net.minecraft.world.level.material.Fluids;
+import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.phys.Vec3;
-import net.minecraft.world.phys.shapes.CollisionContext;
-import net.minecraft.world.phys.shapes.VoxelShape;
+import net.minecraftforge.event.ForgeEventFactory;
 import org.jetbrains.annotations.NotNull;
 
-import javax.annotation.Nullable;
 import java.util.function.Supplier;
 
 @SuppressWarnings("deprecation")
-public class TallAmbientPlantBlock extends AmbientPlantBlock {
+public class TallAmbientPlantBlock extends PrehistoricTallPlantBlock {
 
-    public static final EnumProperty<DoubleBlockHalf> HALF = BlockStateProperties.DOUBLE_BLOCK_HALF;
+    private final Supplier<EntityType<?>> toSpawn;
+    private final int spawnCount;
 
-    protected static final VoxelShape UPPER_SHAPE = Block.box(2.0D, 0.0D, 2.0D, 14.0D, 16.0D, 14.0D);
-    protected static final VoxelShape LOWER_SHAPE = Block.box(2.0D, 0.0D, 2.0D, 14.0D, 16.0D, 14.0D);
-
-    public TallAmbientPlantBlock(Properties properties, Supplier<EntityType<?>> toSpawn) {
-        super(properties, toSpawn);
-        this.registerDefaultState(this.stateDefinition.any().setValue(HALF, DoubleBlockHalf.LOWER));
+    public TallAmbientPlantBlock(Properties properties, Supplier<EntityType<?>> toSpawn, int spawnCount) {
+        super(properties);
+        this.toSpawn = toSpawn;
+        this.spawnCount = spawnCount;
     }
 
     @Override
-    public @NotNull VoxelShape getShape(BlockState state, @NotNull BlockGetter blockGetter, @NotNull BlockPos pos, @NotNull CollisionContext context) {
-        Vec3 offset = state.getOffset(blockGetter, pos);
-        return state.getValue(HALF) == DoubleBlockHalf.LOWER ? LOWER_SHAPE.move(offset.x, offset.y, offset.z) : UPPER_SHAPE.move(offset.x, offset.y, offset.z);
+    public void tick(@NotNull BlockState state, @NotNull ServerLevel level, @NotNull BlockPos pos, @NotNull RandomSource random) {
+        this.spawnEntity(level, pos, state, random);
     }
 
-    @Override
-    public @NotNull BlockState updateShape(BlockState state, Direction direction, BlockState facingState, LevelAccessor level, BlockPos pos, BlockPos facingPos) {
-        DoubleBlockHalf doubleblockhalf = state.getValue(HALF);
-        if (direction.getAxis() != Direction.Axis.Y || doubleblockhalf == DoubleBlockHalf.LOWER != (direction == Direction.UP) || facingState.is(this) && facingState.getValue(HALF) != doubleblockhalf) {
-            return doubleblockhalf == DoubleBlockHalf.LOWER && direction == Direction.DOWN && !state.canSurvive(level, pos) ? Blocks.AIR.defaultBlockState() : super.updateShape(state, direction, facingState, level, pos, facingPos);
-        } else {
-            return Blocks.AIR.defaultBlockState();
-        }
-    }
-
-    @Nullable
-    public BlockState getStateForPlacement(BlockPlaceContext context) {
-        BlockPos blockpos = context.getClickedPos();
-        Level level = context.getLevel();
-        return blockpos.getY() < level.getMaxBuildHeight() - 1 && level.getBlockState(blockpos.above()).canBeReplaced(context) ? super.getStateForPlacement(context) : null;
-    }
-
-    @Override
-    public void setPlacedBy(Level level, @NotNull BlockPos pos, @NotNull BlockState state, LivingEntity pPlacer, @NotNull ItemStack stack) {
-        super.setPlacedBy(level, pos, state, pPlacer, stack);
-        BlockPos blockpos = pos.above();
-        level.setBlock(blockpos, copyWaterloggedFrom(level, blockpos, this.defaultBlockState().setValue(HALF, DoubleBlockHalf.UPPER)), 3);
-    }
-
-    @Override
-    public boolean canSurvive(BlockState state, LevelReader level, BlockPos pos) {
+    public void spawnEntity(ServerLevel level, BlockPos pos, BlockState state, RandomSource random) {
         if (state.getValue(HALF) != DoubleBlockHalf.UPPER) {
-            return super.canSurvive(state, level, pos);
-        } else {
-            BlockState blockstate = level.getBlockState(pos.below());
-            if (state.getBlock() != this) return super.canSurvive(state, level, pos);
-            return blockstate.is(this) && blockstate.getValue(HALF) == DoubleBlockHalf.LOWER;
-        }
-    }
-
-    public static void placeAt(LevelAccessor level, BlockState state, BlockPos pos, int flags) {
-        BlockPos blockpos = pos.above();
-        level.setBlock(pos, copyWaterloggedFrom(level, pos, state.setValue(HALF, DoubleBlockHalf.LOWER)), flags);
-        level.setBlock(blockpos, copyWaterloggedFrom(level, blockpos, state.setValue(HALF, DoubleBlockHalf.UPPER)), flags);
-    }
-
-    public static BlockState copyWaterloggedFrom(LevelReader level, BlockPos pos, BlockState state) {
-        return state.hasProperty(BlockStateProperties.WATERLOGGED) ? state.setValue(BlockStateProperties.WATERLOGGED, level.isWaterAt(pos)) : state;
-    }
-
-    @Override
-    public void playerWillDestroy(Level level, @NotNull BlockPos pos, @NotNull BlockState state, @NotNull Player player) {
-        if (!level.isClientSide) {
-            if (player.isCreative()) {
-                preventCreativeDropFromBottomPart(level, pos, state, player);
-            } else {
-                dropResources(state, level, pos, null, player, player.getMainHandItem());
+            level.playSound(null, pos, SoundEvents.ITEM_PICKUP, SoundSource.BLOCKS, 0.1F, 0.8F + random.nextFloat() * 0.2F);
+            for (int j = 0; j < this.spawnCount; j++) {
+                Entity entity = this.toSpawn.get().create(level);
+                Vec3 vec3 = pos.getCenter();
+                if (entity instanceof AmbientMob mob) {
+                    mob.setShouldBeRestricted(true);
+                    entity.moveTo(vec3.x(), vec3.y(), vec3.z(), Mth.wrapDegrees(level.random.nextFloat() * 360.0F), 0.0F);
+                    level.addFreshEntity(entity);
+                    ForgeEventFactory.onFinalizeSpawn(mob, level, level.getCurrentDifficultyAt(pos), MobSpawnType.NATURAL, null, null);
+                }
             }
-        }
-        super.playerWillDestroy(level, pos, state, player);
-    }
-
-    @Override
-    public void playerDestroy(@NotNull Level level, @NotNull Player player, @NotNull BlockPos pos, @NotNull BlockState state, @Nullable BlockEntity blockEntity, ItemStack stack) {
-        super.playerDestroy(level, player, pos, Blocks.AIR.defaultBlockState(), blockEntity, stack);
-    }
-
-    protected static void preventCreativeDropFromBottomPart(Level level, BlockPos pos, BlockState state, Player player) {
-        DoubleBlockHalf doubleblockhalf = state.getValue(HALF);
-        if (doubleblockhalf == DoubleBlockHalf.UPPER) {
-            BlockPos blockpos = pos.below();
-            BlockState blockstate = level.getBlockState(blockpos);
-            if (blockstate.is(state.getBlock()) && blockstate.getValue(HALF) == DoubleBlockHalf.LOWER) {
-                BlockState blockstate1 = blockstate.getFluidState().is(Fluids.WATER) ? Blocks.WATER.defaultBlockState() : Blocks.AIR.defaultBlockState();
-                level.setBlock(blockpos, blockstate1, 35);
-                level.levelEvent(player, 2001, blockpos, Block.getId(blockstate));
-            }
+            level.scheduleTick(pos, this, 2600 + level.getRandom().nextInt(1200));
         }
     }
 
     @Override
-    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> pBuilder) {
-        pBuilder.add(HALF);
-    }
-
-    @Override
-    public long getSeed(BlockState state, BlockPos pos) {
-        return Mth.getSeed(pos.getX(), pos.below(state.getValue(HALF) == DoubleBlockHalf.LOWER ? 0 : 1).getY(), pos.getZ());
+    public void onPlace(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull BlockState blockState, boolean b) {
+        if (state.getValue(HALF) != DoubleBlockHalf.UPPER) {
+            level.gameEvent(GameEvent.BLOCK_PLACE, pos, GameEvent.Context.of(state));
+            level.scheduleTick(pos, this, 2600 + level.getRandom().nextInt(1200));
+        }
     }
 }