+++ src/main/java/com/barlinc/unusual_prehistory/client/models/entity/mob/update_5/ambient/ZhangsolvaModel.java
@@ -1,8 +1,8 @@
-package com.barlinc.unusual_prehistory.client.models.entity.mob.future.ambient;
+package com.barlinc.unusual_prehistory.client.models.entity.mob.update_5.ambient;
 
 import com.barlinc.unusual_prehistory.client.animations.entity.mob.future.ambient.ZhangsolvaAnimations;
 import com.barlinc.unusual_prehistory.client.models.entity.UP2Model;
-import com.barlinc.unusual_prehistory.entity.mob.future.ambient.Zhangsolva;
+import com.barlinc.unusual_prehistory.entity.mob.update_5.ambient.Zhangsolva;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.model.geom.PartPose;
 import net.minecraft.client.model.geom.builders.*;