+++ src/main/java/com/barlinc/unusual_prehistory/datagen/UP2DamageTypeTagProvider.java
@@ -14,7 +14,8 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.CompletableFuture;
 
-import static com.barlinc.unusual_prehistory.registry.UP2DamageTypes.*;
+import static com.barlinc.unusual_prehistory.registry.UP2DamageTypes.EXECUTE;
+import static com.barlinc.unusual_prehistory.registry.UP2DamageTypes.GRUG;
 
 public class UP2DamageTypeTagProvider extends TagsProvider<DamageType> {