+++ src/main/java/com/barlinc/unusual_prehistory/entity/mob/update_5/ambient/Zhangsolva.java
@@ -1,6 +1,7 @@
-package com.barlinc.unusual_prehistory.entity.mob.future.ambient;
+package com.barlinc.unusual_prehistory.entity.mob.update_5.ambient;
 
 import com.barlinc.unusual_prehistory.UnusualPrehistory2;
+import com.barlinc.unusual_prehistory.entity.ai.goals.AmbientMoveToRestrictionGoal;
 import com.barlinc.unusual_prehistory.entity.ai.navigation.NoSpinFlyingPathNavigation;
 import com.barlinc.unusual_prehistory.entity.mob.base.AmbientMob;
 import com.barlinc.unusual_prehistory.registry.UP2SoundEvents;
@@ -8,6 +9,7 @@ import com.barlinc.unusual_prehistory.utils.SmoothAnimationState;
 import net.minecraft.core.BlockPos;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.MobType;
@@ -28,12 +30,12 @@ import org.jetbrains.annotations.NotNull;
 import javax.annotation.Nullable;
 import java.util.EnumSet;
 
-public class Delitzschala extends AmbientMob {
+public class Zhangsolva extends AmbientMob {
 
     public final SmoothAnimationState idleAnimationState = new SmoothAnimationState(1.0F);
     public final SmoothAnimationState flyAnimationState = new SmoothAnimationState(1.0F);
 
-    public Delitzschala(EntityType<? extends AmbientMob> entityType, Level level) {
+    public Zhangsolva(EntityType<? extends AmbientMob> entityType, Level level) {
         super(entityType, level);
         this.moveControl = new FlyingMoveControl(this, 20, false);
     }
@@ -42,14 +44,15 @@ public class Delitzschala extends AmbientMob {
         return Mob.createMobAttributes()
                 .add(Attributes.MAX_HEALTH, 1.0D)
                 .add(Attributes.MOVEMENT_SPEED, 0.01F)
-                .add(Attributes.FLYING_SPEED, 0.7F);
+                .add(Attributes.FLYING_SPEED, 0.5F);
     }
 
     @Override
     protected void registerGoals() {
         this.goalSelector.addGoal(0, new FloatGoal(this));
-        this.goalSelector.addGoal(1, new DelitzschalaWanderAroundGoal(this));
-        this.goalSelector.addGoal(2, new DelitzschalaLookAroundGoal(this));
+        this.goalSelector.addGoal(1, new AmbientMoveToRestrictionGoal(this, 1.0D));
+        this.goalSelector.addGoal(2, new ZhangsolvaWanderAroundGoal(this));
+        this.goalSelector.addGoal(3, new ZhangsolvaLookAroundGoal(this));
     }
 
     @Override
@@ -77,7 +80,7 @@ public class Delitzschala extends AmbientMob {
     }
 
     @Override
-    public void remove(@NotNull RemovalReason removalReason) {
+    public void remove(Entity.@NotNull RemovalReason removalReason) {
         UnusualPrehistory2.PROXY.clearSoundCacheFor(this);
         super.remove(removalReason);
     }
@@ -86,7 +89,7 @@ public class Delitzschala extends AmbientMob {
     public void tick() {
         super.tick();
 
-        if (!this.onGround() && this.getDeltaMovement().y < 0.0) {
+        if (!this.onGround() && this.getDeltaMovement().y < 0.0 && this.isAlive()) {
             this.setDeltaMovement(this.getDeltaMovement().multiply(1.0, 0.6, 1.0));
         }
 
@@ -111,58 +114,58 @@ public class Delitzschala extends AmbientMob {
         return UP2SoundEvents.BUG_DEATH.get();
     }
 
-    private static class DelitzschalaWanderAroundGoal extends Goal {
+    private static class ZhangsolvaWanderAroundGoal extends Goal {
 
-        private final Delitzschala delitzschala;
+        private final Zhangsolva zhangsolva;
 
-        public DelitzschalaWanderAroundGoal(Delitzschala delitzschala) {
-            this.delitzschala = delitzschala;
-            this.setFlags(EnumSet.of(Flag.MOVE));
+        public ZhangsolvaWanderAroundGoal(Zhangsolva zhangsolva) {
+            this.zhangsolva = zhangsolva;
+            this.setFlags(EnumSet.of(Goal.Flag.MOVE));
         }
 
         @Override
         public boolean canUse() {
-            return delitzschala.navigation.isDone() && delitzschala.random.nextInt(10) == 0;
+            return zhangsolva.navigation.isDone() && zhangsolva.random.nextInt(20) == 0;
         }
 
         @Override
         public boolean canContinueToUse() {
-            return delitzschala.navigation.isInProgress();
+            return zhangsolva.navigation.isInProgress();
         }
 
         @Override
         public void start() {
             Vec3 vec3d = this.getRandomLocation();
             if (vec3d != null) {
-                this.delitzschala.navigation.moveTo(delitzschala.navigation.createPath(BlockPos.containing(vec3d), 1), 1.0);
+                this.zhangsolva.navigation.moveTo(zhangsolva.navigation.createPath(BlockPos.containing(vec3d), 1), 1.0);
             }
         }
 
         @Nullable
         private Vec3 getRandomLocation() {
-            Vec3 vec3d2 = delitzschala.getViewVector(0.0F);
-            Vec3 vec3d3 = HoverRandomPos.getPos(delitzschala, 10, 3, vec3d2.x, vec3d2.z, 1.5707964F, 3, 1);
-            return vec3d3 != null ? vec3d3 : AirAndWaterRandomPos.getPos(delitzschala, 10, 2, -2, vec3d2.x, vec3d2.z, 1.5707963705062866);
+            Vec3 vec3d2 = zhangsolva.getViewVector(0.0F);
+            Vec3 vec3d3 = HoverRandomPos.getPos(zhangsolva, 10, 3, vec3d2.x, vec3d2.z, 1.5707964F, 3, 1);
+            return vec3d3 != null ? vec3d3 : AirAndWaterRandomPos.getPos(zhangsolva, 10, 2, -2, vec3d2.x, vec3d2.z, 1.5707963705062866);
         }
     }
 
-    private class DelitzschalaLookAroundGoal extends RandomLookAroundGoal {
+    private class ZhangsolvaLookAroundGoal extends RandomLookAroundGoal {
 
-        private final Delitzschala delitzschala;
+        private final Zhangsolva zhangsolva;
 
-        public DelitzschalaLookAroundGoal(Delitzschala delitzschala) {
-            super(Delitzschala.this);
-            this.delitzschala = delitzschala;
+        public ZhangsolvaLookAroundGoal(Zhangsolva zhangsolva) {
+            super(Zhangsolva.this);
+            this.zhangsolva = zhangsolva;
         }
 
         @Override
         public boolean canUse() {
-            return delitzschala.onGround() && super.canUse();
+            return zhangsolva.onGround() && super.canUse();
         }
 
         @Override
         public boolean canContinueToUse() {
-            return delitzschala.onGround() && super.canContinueToUse();
+            return zhangsolva.onGround() && super.canContinueToUse();
         }
     }
 }