+++ src/main/java/com/barlinc/unusual_prehistory/events/ClientModEvents.java
@@ -1,8 +1,6 @@
 package com.barlinc.unusual_prehistory.events;
 
 import com.barlinc.unusual_prehistory.UnusualPrehistory2;
-import com.barlinc.unusual_prehistory.client.models.entity.mob.future.ambient.DelitzschalaModel;
-import com.barlinc.unusual_prehistory.client.models.entity.mob.future.ambient.ZhangsolvaModel;
 import com.barlinc.unusual_prehistory.client.models.entity.mob.update_1.*;
 import com.barlinc.unusual_prehistory.client.models.entity.mob.update_1.jawless_fish.*;
 import com.barlinc.unusual_prehistory.client.models.entity.mob.update_2.OnchopristisModel;
@@ -12,11 +10,11 @@ import com.barlinc.unusual_prehistory.client.models.entity.mob.update_3.Tartuost
 import com.barlinc.unusual_prehistory.client.models.entity.mob.update_4.*;
 import com.barlinc.unusual_prehistory.client.models.entity.mob.update_4.lobe_finned_fish.*;
 import com.barlinc.unusual_prehistory.client.models.entity.mob.update_5.*;
+import com.barlinc.unusual_prehistory.client.models.entity.mob.update_5.ambient.DelitzschalaModel;
+import com.barlinc.unusual_prehistory.client.models.entity.mob.update_5.ambient.ZhangsolvaModel;
 import com.barlinc.unusual_prehistory.client.particles.*;
 import com.barlinc.unusual_prehistory.client.renderer.entity.UP2BoatRenderer;
 import com.barlinc.unusual_prehistory.client.renderer.entity.UP2ChestBoatRenderer;
-import com.barlinc.unusual_prehistory.client.renderer.entity.mob.future.ambient.DelitzschalaRenderer;
-import com.barlinc.unusual_prehistory.client.renderer.entity.mob.future.ambient.ZhangsolvaRenderer;
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_1.*;
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_2.OnchopristisRenderer;
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_3.LivingOozeRenderer;
@@ -24,6 +22,8 @@ import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_3.Metrio
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_3.TartuosteusRenderer;
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_4.*;
 import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_5.*;
+import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_5.ambient.DelitzschalaRenderer;
+import com.barlinc.unusual_prehistory.client.renderer.entity.mob.update_5.ambient.ZhangsolvaRenderer;
 import com.barlinc.unusual_prehistory.registry.*;
 import com.barlinc.unusual_prehistory.screens.TransmogrifierScreen;
 import net.minecraft.client.gui.screens.MenuScreens;