+++ src/main/java/com/barlinc/unusual_prehistory/registry/UP2BlockEntities.java
@@ -1,7 +1,6 @@
 package com.barlinc.unusual_prehistory.registry;
 
 import com.barlinc.unusual_prehistory.UnusualPrehistory2;
-import com.barlinc.unusual_prehistory.blocks.entity.AmbientPlantBlockEntity;
 import com.barlinc.unusual_prehistory.blocks.entity.ExtraDataBlockEntity;
 import com.barlinc.unusual_prehistory.blocks.entity.TransmogrifierBlockEntity;
 import net.minecraft.world.level.block.Block;
@@ -18,9 +17,5 @@ public class UP2BlockEntities {
 
     public static final RegistryObject<BlockEntityType<TransmogrifierBlockEntity>> TRANSMOGRIFIER = BLOCK_ENTITIES.register("transmogrifier", () -> BlockEntityType.Builder.of(TransmogrifierBlockEntity::new, UP2Blocks.TRANSMOGRIFIER.get()).build(null));
     public static final RegistryObject<BlockEntityType<ExtraDataBlockEntity>> EXTRA_DATA_BLOCK_ENTITY = BLOCK_ENTITIES.register("extra_data_block_entity", () -> BlockEntityType.Builder.of(ExtraDataBlockEntity::new, UP2Blocks.EGG_BLOCKS.stream().map(Supplier::get).toArray(Block[]::new)).build(null));
-
-    public static final RegistryObject<BlockEntityType<AmbientPlantBlockEntity>> AMBIENT_PLANT = BLOCK_ENTITIES.register("ambient_plant", () -> BlockEntityType.Builder.of(AmbientPlantBlockEntity::new,
-            UP2Blocks.ZHANGSOLVA_BLOOM.get()
-    ).build(null));
 
 }